+++ src/raft/raft.go
@@ -28,47 +28,6 @@ import (
 	"6.5840/labrpc"
 )
 
-// as each Raft peer becomes aware that successive log entries are
-// committed, the peer should send an ApplyMsg to the service (or
-// tester) on the same server, via the applyCh passed to Make(). set
-// CommandValid to true to indicate that the ApplyMsg contains a newly
-// committed log entry.
-//
-// in part 2D you'll want to send other kinds of messages (e.g.,
-// snapshots) on the applyCh, but set CommandValid to false for these
-// other uses.
-type ApplyMsg struct {
-	CommandValid bool
-	Command      interface{}
-	CommandIndex int
-
-	// For 2D:
-	SnapshotValid bool
-	Snapshot      []byte
-	SnapshotTerm  int
-	SnapshotIndex int
-}
-
-// A Go object implementing a single Raft peer.
-type Raft struct {
-	mu              sync.Mutex          // Lock to protect shared access to this peer's state
-	peers           []*labrpc.ClientEnd // RPC end points of all peers
-	persister       *Persister          // Object to hold this peer's persisted state
-	me              int                 // this peer's index into peers[]
-	dead            int32               // set by Kill()
-	role            RaftRole
-	lastHeartBeat   time.Time
-	electionTrigger chan bool
-
-	// Your data here (2A, 2B, 2C).
-	// Look at the paper's Figure 2 for a description of what
-	// state a Raft server must maintain.
-
-	// for leader election
-	currentTerm int
-	votedFor    int
-}
-
 const (
 	heartBeatIntervalMS  = 100
 	electionTimeoutMinMS = 300
@@ -368,15 +327,20 @@ func Make(peers []*labrpc.ClientEnd, me int,
 		votedFor:        -1,
 		lastHeartBeat:   time.Now(),
 		electionTrigger: make(chan bool),
+		applyCh:         applyCh,
 		currentTerm:     0,
+		log:             []LogEntry{{-1, -1} /*dummy head*/},
+		commitIdx:       0,
+		lastApplied:     0,
 	}
+	rf.applyCond = sync.NewCond(&rf.mu)
+
 	// Your initialization code here (2A, 2B, 2C).
 
-	// timer for leader election
-
-	// initialize from state persisted before a crash
+	// initialize from state persisted before a crash.
 	rf.readPersist(persister.ReadRaftState())
 
+	// run an election when timeout ticker goes off.
 	go rf.timeoutTicker(rf.electionTrigger)
 	go rf.runElection(rf.electionTrigger)
