+++ src/raft/type.go
@@ -1,5 +1,12 @@
 package raft
 
+import (
+	"sync"
+	"time"
+
+	"6.5840/labrpc"
+)
+
 type RaftRole string
 
 const (
@@ -11,6 +18,59 @@ const (
 	RaftRPCRequestVote   = "Raft.RequestVote"
 	RaftRPCAppendENtries = "Raft.AppendEntries"
 )
+
+// as each Raft peer becomes aware that successive log entries are
+// committed, the peer should send an ApplyMsg to the service (or
+// tester) on the same server, via the applyCh passed to Make(). set
+// CommandValid to true to indicate that the ApplyMsg contains a newly
+// committed log entry.
+//
+// in part 2D you'll want to send other kinds of messages (e.g.,
+// snapshots) on the applyCh, but set CommandValid to false for these
+// other uses.
+type ApplyMsg struct {
+	CommandValid bool
+	Command      interface{}
+	CommandIndex int
+
+	// For 2D:
+	SnapshotValid bool
+	Snapshot      []byte
+	SnapshotTerm  int
+	SnapshotIndex int
+}
+
+type LogEntry struct {
+	data interface{}
+	term int
+}
+
+// A Go object implementing a single Raft peer.
+type Raft struct {
+	mu              sync.Mutex          // Lock to protect shared access to this peer's state
+	peers           []*labrpc.ClientEnd // RPC end points of all peers
+	persister       *Persister          // Object to hold this peer's persisted state
+	me              int                 // this peer's index into peers[]
+	dead            int32               // set by Kill()
+	role            RaftRole
+	lastHeartBeat   time.Time
+	electionTrigger chan bool
+	applyCh         chan ApplyMsg // for test only
+
+	// Your data here (2A, 2B, 2C).
+	// Look at the paper's Figure 2 for a description of what
+	// state a Raft server must maintain.
+	currentTerm int
+	votedFor    int
+	log         []LogEntry
+	commitIdx   int
+	lastApplied int
+	applyCond   *sync.Cond
+
+	// only for leader
+	nextIdx  []int
+	matchIdx []int
+}
 
 type RequestVoteArgs struct {
 	Term         int
@@ -25,8 +85,12 @@ type RequestVoteReply struct {
 }
 
 type AppendEntriesArgs struct {
-	Term     int // leader's term
-	LeaderId int // Follower can redirect to leader
+	Term            int // leader's term
+	LeaderId        int // Follower can redirect to leader
+	PrevLogIdx      int
+	PrevLogTerm     int
+	Entries         []LogEntry // empty for heartbeat, may send more than one for efficiency
+	LeaderCommitIdx int
 }
 
 type AppendEntriesReply struct {
